add ethers support
