+++ examples/with-biconomy-aa/.env.local.example
@@ -4,3 +4,5 @@ BASE_URL="https://api.turnkey.com"
 ORGANIZATION_ID="<Turnkey organization ID>"
 SIGN_WITH="<Turnkey Wallet Account Address, Private Key Address, or Private Key ID>" # if blank, we will create a wallet for you
 INFURA_KEY="<Infura API Key>"
+BICONOMY_BUNDLER_URL="<Biconomy Bundler URL>" # see https://docs.biconomy.io/dashboard#bundler-url
+BICONOMY_PAYMASTER_API_KEY="<Biconomy Paymaster API Key>" # see https://docs.biconomy.io/dashboard/paymaster
