+++ examples/with-biconomy-aa/package.json
@@ -1,23 +1,21 @@
 {
-  "name": "@turnkey/example-with-ethers",
-  "version": "0.1.3",
+  "name": "@turnkey/example-with-biconomy-aa",
+  "version": "0.1.0",
   "private": true,
   "scripts": {
-    "start": "tsx src/index.ts",
-    "start-advanced": "tsx src/advanced.ts",
-    "start-sepolia-legacy": "tsx src/sepoliaLegacyTx.ts",
+    "start-ethers": "tsx src/ethers.ts",
     "clean": "rimraf ./dist ./.cache",
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
     "@biconomy/account": "^4.5.6",
-    "@turnkey/api-key-stamper": "workspace:*",
     "@turnkey/ethers": "workspace:*",
-    "@turnkey/http": "workspace:*",
+    "@turnkey/viem": "workspace:*",
     "@turnkey/sdk-server": "workspace:*",
     "dotenv": "^16.0.3",
     "ethers": "^6.10.0",
-    "prompts": "^2.4.2"
+    "prompts": "^2.4.2",
+    "viem": "^1.16.6"
   },
   "devDependencies": {
     "@types/prompts": "^2.4.2"
