+++ examples/with-biconomy-aa/src/createNewWallet.ts
@@ -1,9 +1,5 @@
-import {
-  TurnkeyClient,
-  createActivityPoller,
-  TurnkeyActivityError,
-} from "@turnkey/http";
-import { ApiKeyStamper } from "@turnkey/api-key-stamper";
+import { Turnkey as TurnkeySDKServer } from "@turnkey/sdk-server";
+
 import * as crypto from "crypto";
 import { refineNonNull } from "./util";
 
@@ -13,24 +9,16 @@ export async function createNewWallet() {
   const walletName = `ETH Wallet ${crypto.randomBytes(2).toString("hex")}`;
 
   try {
-    const turnkeyClient = new TurnkeyClient(
-      { baseUrl: process.env.BASE_URL! },
-      new ApiKeyStamper({
-        apiPublicKey: process.env.API_PUBLIC_KEY!,
-        apiPrivateKey: process.env.API_PRIVATE_KEY!,
-      })
-    );
-
-    const activityPoller = createActivityPoller({
-      client: turnkeyClient,
-      requestFn: turnkeyClient.createWallet,
+    const turnkeyClient = new TurnkeySDKServer({
+      apiBaseUrl: "https://api.turnkey.com",
+      apiPublicKey: process.env.API_PUBLIC_KEY!,
+      apiPrivateKey: process.env.API_PRIVATE_KEY!,
+      defaultOrganizationId: process.env.ORGANIZATION_ID!,
     });
 
-    const completedActivity = await activityPoller({
-      type: "ACTIVITY_TYPE_CREATE_WALLET",
-      timestampMs: String(Date.now()),
-      organizationId: process.env.ORGANIZATION_ID!,
-      parameters: {
+    const { walletId, addresses } = await turnkeyClient
+      .apiClient()
+      .createWallet({
         walletName,
         accounts: [
           {
@@ -40,33 +28,23 @@ export async function createNewWallet() {
             addressFormat: "ADDRESS_FORMAT_ETHEREUM",
           },
         ],
-      },
-    });
+      });
 
-    const wallet = refineNonNull(completedActivity.result.createWalletResult);
-    const walletId = refineNonNull(wallet.walletId);
-    const address = refineNonNull(wallet.addresses[0]);
+    const newWalletId = refineNonNull(walletId);
+    const address = refineNonNull(addresses[0]);
 
     // Success!
     console.log(
       [
         `New Ethereum wallet created!`,
         `- Name: ${walletName}`,
-        `- Wallet ID: ${walletId}`,
+        `- Wallet ID: ${newWalletId}`,
         `- Address: ${address}`,
         ``,
         "Now you can take the address, put it in `.env.local` (`SIGN_WITH=<address>`), then re-run the script.",
       ].join("\n")
     );
   } catch (error) {
-    // If needed, you can read from `TurnkeyActivityError` to find out why the activity didn't succeed
-    if (error instanceof TurnkeyActivityError) {
-      throw error;
-    }
-
-    throw new TurnkeyActivityError({
-      message: "Failed to create a new Ethereum wallet",
-      cause: error as Error,
-    });
+    throw new Error("Failed to create a new Ethereum wallet: " + error);
   }
 }
